+++ Dtos/DtoResponse.cs
@@ -4,7 +4,7 @@ namespace projeto1
     public class DtoResponse
     {
         public string Nome {get; set;} = string.Empty;
-               public int Avaliacao {get; set;}
+               public decimal Avaliacao {get; set;}
                 public StatusJogo Status {get; set;}
